+++ docs/src/release_notes.rst
@@ -3,6 +3,14 @@
 
 Release notes
 =============
+Unreleased
+----------
+Added
+*****
+- :ref:`auth` - Add :attr:`streaming <scope.streaming>` and
+  :attr:`app-remote-control <scope.app_remote_control>`
+  as extra scopes (:issue:`237`)
+
 3.4.2 (2020-12-14)
 ------------------
 Fixed